format test files
